backend/Web-Shop-Backend/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=dev.txt dev.in
#
asgiref==3.8.1
    # via
    #   -r base.txt
    #   django
    #   django-cors-headers
    #   django-stubs
asttokens==3.0.0
    # via
    #   -r base.txt
    #   stack-data
certifi==2024.12.14
    # via
    #   -r base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r base.txt
    #   cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via
    #   -r base.txt
    #   requests
coverage[toml]==7.6.10
    # via pytest-cov
cryptography==44.0.0
    # via
    #   -r base.txt
    #   social-auth-core
decorator==5.1.1
    # via
    #   -r base.txt
    #   ipython
defusedxml==0.8.0rc2
    # via
    #   -r base.txt
    #   python3-openid
    #   social-auth-core
distlib==0.3.9
    # via virtualenv
django==5.1.4
    # via
    #   -r base.txt
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-jazzmin
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-yasg
    #   social-auth-app-django
django-cors-headers==4.6.0
    # via -r base.txt
django-extensions==3.2.3
    # via -r base.txt
django-filter==24.3
    # via -r base.txt
django-jazzmin==3.0.1
    # via -r base.txt
django-stubs==5.1.1
    # via -r dev.in
django-stubs-ext==5.1.1
    # via django-stubs
djangorestframework==3.15.2
    # via
    #   -r base.txt
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.4.0
    # via
    #   -r base.txt
    #   djoser
djoser==2.3.1
    # via -r base.txt
drf-yasg==1.21.8
    # via -r base.txt
environs==14.1.0
    # via -r base.txt
execnet==2.1.1
    # via pytest-xdist
executing==2.1.0
    # via
    #   -r base.txt
    #   stack-data
factory-boy==3.3.1
    # via -r dev.in
faker==33.3.1
    # via factory-boy
filelock==3.16.1
    # via virtualenv
identify==2.6.5
    # via pre-commit
idna==3.10
    # via
    #   -r base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r base.txt
    #   drf-yasg
iniconfig==2.0.0
    # via pytest
ipython==8.31.0
    # via -r base.txt
jedi==0.19.2
    # via
    #   -r base.txt
    #   ipython
marshmallow==3.25.1
    # via
    #   -r base.txt
    #   environs
matchlib==0.2.3
    # via -r dev.in
matplotlib-inline==0.1.7
    # via
    #   -r base.txt
    #   ipython
mypy==1.14.1
    # via -r dev.in
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
oauthlib==3.2.2
    # via
    #   -r base.txt
    #   requests-oauthlib
    #   social-auth-core
packaging==24.2
    # via
    #   -r base.txt
    #   drf-yasg
    #   marshmallow
    #   pytest
    #   pytest-sugar
parso==0.8.4
    # via
    #   -r base.txt
    #   jedi
pexpect==4.9.0
    # via
    #   -r base.txt
    #   ipython
pillow==11.1.0
    # via -r base.txt
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via -r dev.in
prompt-toolkit==3.0.48
    # via
    #   -r base.txt
    #   ipython
psycopg2-binary==2.9.10
    # via -r base.txt
ptyprocess==0.7.0
    # via
    #   -r base.txt
    #   pexpect
pure-eval==0.2.3
    # via
    #   -r base.txt
    #   stack-data
pycparser==2.22
    # via
    #   -r base.txt
    #   cffi
pygments==2.19.1
    # via
    #   -r base.txt
    #   ipython
pyjwt==2.10.1
    # via
    #   -r base.txt
    #   djangorestframework-simplejwt
    #   social-auth-core
pytest==8.3.4
    # via
    #   -r dev.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-django
    #   pytest-sugar
    #   pytest-xdist
pytest-asyncio==0.25.2
    # via -r dev.in
pytest-cov==6.0.0
    # via -r dev.in
pytest-django==4.9.0
    # via -r dev.in
pytest-sugar==1.0.0
    # via -r dev.in
pytest-xdist==3.6.1
    # via -r dev.in
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   time-machine
python-dotenv==1.0.1
    # via
    #   -r base.txt
    #   environs
python3-openid==3.2.0
    # via
    #   -r base.txt
    #   social-auth-core
pytz==2024.2
    # via
    #   -r base.txt
    #   drf-yasg
pyyaml==6.0.2
    # via
    #   -r base.txt
    #   drf-yasg
    #   pre-commit
requests==2.32.3
    # via
    #   -r base.txt
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via
    #   -r base.txt
    #   social-auth-core
ruff==0.9.1
    # via -r dev.in
six==1.17.0
    # via python-dateutil
social-auth-app-django==5.4.2
    # via
    #   -r base.txt
    #   djoser
social-auth-core==4.5.4
    # via
    #   -r base.txt
    #   social-auth-app-django
sqlparse==0.5.3
    # via
    #   -r base.txt
    #   django
stack-data==0.6.3
    # via
    #   -r base.txt
    #   ipython
termcolor==2.5.0
    # via pytest-sugar
time-machine==2.16.0
    # via -r dev.in
traitlets==5.14.3
    # via
    #   -r base.txt
    #   ipython
    #   matplotlib-inline
types-pyyaml==6.0.12.20241230
    # via django-stubs
typing-extensions==4.12.2
    # via
    #   -r base.txt
    #   django-stubs
    #   django-stubs-ext
    #   faker
    #   ipython
    #   mypy
uritemplate==4.1.1
    # via
    #   -r base.txt
    #   drf-yasg
urllib3==2.3.0
    # via
    #   -r base.txt
    #   requests
virtualenv==20.28.1
    # via pre-commit
wcwidth==0.2.13
    # via
    #   -r base.txt
    #   prompt-toolkit
